requirements.txt
asgiref==3.7.2
Babel==2.8.0
bcrypt==4.0.1
blinker==1.6.3
cachelib==0.10.2
cffi==1.16.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cs50==9.3.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Flask==3.0.0
Flask-Session==0.5.0
greenlet==3.0.1
gunicorn==21.2.0
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
more-itertools==8.10.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
paramiko==3.3.1
pip-install==1.3.5
protobuf==4.21.12
pycparser==2.21
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
pytz==2022.1
PyYAML==5.4.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sshtunnel==0.4.0
systemd-python==234
termcolor==2.3.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
